Fixed gcc 4.8.1 -Wformat warnings on windows.
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=67072

+++ dbus/dbus-sysdeps-win.h
@@ -44,7 +44,7 @@ const char* _dbus_win_error_from_last_error (void);
 
 void _dbus_win_startup_winsock (void);
 void _dbus_win_warn_win_error  (const char *message,
-                                int         code);
+                                unsigned long code);
                                 
 char * _dbus_win_error_string (int error_number);
 void _dbus_win_free_error_string (char *string);
